+++ src/main/webapp/app/admin/salon/list/salon.component.ts
@@ -1,26 +1,28 @@
 import {Component, inject, OnInit} from '@angular/core';
 import {ActivatedRoute, Router, RouterModule} from '@angular/router';
 import {filter, switchMap, tap} from 'rxjs';
-import {NgbModal} from '@ng-bootstrap/ng-bootstrap';
 
 import SharedModule from 'app/shared/shared.module';
 import {FormatMediumDatePipe} from 'app/shared/date';
 import {FormsModule} from '@angular/forms';
-import {ITEM_DELETED_EVENT} from 'app/config/navigation.constants';
 import {ISalon} from '../model/salon.interface';
 import {SalonService} from '../service/salon.service';
-import {DeleteDialogComponent} from '../../../shared/delete-dialog/delete-dialog.component';
 import {finalize} from 'rxjs/operators';
 import {ButtonBoxComponent} from '../../../shared/components/button-box/button-box.component';
-import {LinkBoxComponent} from '../../../shared/components/link-box/link-box.component';
-import {PaginationComponent} from '../../../shared/pagination/pagination.component';
-import {PaginationEvent} from '../../../shared/pagination/pagination-event.interface';
 import {AlertErrorComponent} from "../../../shared/alert/alert-error.component";
 import {AlertComponent} from "../../../shared/alert/alert.component";
-import {ProgressSpinner} from "primeng/progressspinner";
+import {ConfirmDialogService} from "../../../shared/delete-dialog/confirm-dialog.service";
+import {ConfirmPopup} from "primeng/confirmpopup";
+import {Toast} from "primeng/toast";
+import {TableModule} from "primeng/table";
+import {ContentPageComponent} from "../../../shared/components/content-page/content-page.component";
+import {CardComponent} from "../../../shared/components/card/card.component";
+import {IconField} from "primeng/iconfield";
+import {InputIcon} from "primeng/inputicon";
+import {InputText} from "primeng/inputtext";
 
 @Component({
-    selector: 'jhi-salon',
+    selector: 'app-salon',
     templateUrl: './salon.component.html',
     imports: [
         RouterModule,
@@ -28,20 +30,24 @@ import {ProgressSpinner} from "primeng/progressspinner";
         SharedModule,
         FormatMediumDatePipe,
         ButtonBoxComponent,
-        LinkBoxComponent,
-        PaginationComponent,
         AlertErrorComponent,
         AlertComponent,
-        ProgressSpinner,
+        ConfirmPopup,
+        Toast,
+        TableModule,
+        ContentPageComponent,
+        CardComponent,
+        IconField,
+        InputIcon,
+        InputText,
     ]
 })
 export class SalonComponent implements OnInit {
     public router = inject(Router);
     protected salonService = inject(SalonService);
     protected activatedRoute = inject(ActivatedRoute);
-    protected modalService = inject(NgbModal);
+    protected confirmDialogService = inject(ConfirmDialogService);
 
-    salonsPaginated: ISalon[] = [];
     salons: ISalon[] = [];
     isLoading = false;
 
@@ -51,21 +57,12 @@ export class SalonComponent implements OnInit {
         }
     }
 
-    delete(salon: ISalon): void {
-        const modalRef = this.modalService.open(DeleteDialogComponent, {
-            size: 'lg',
-            backdrop: 'static',
-        });
-        modalRef.componentInstance.translateKey = 'salon.delete.question';
-        modalRef.componentInstance.translateValues = {id: salon.place};
-
-        modalRef.closed
-            .pipe(
-                filter((reason) => reason === ITEM_DELETED_EVENT),
-                switchMap(() => this.salonService.delete(salon.id)),
-                tap(() => this.load()), // Recharge les données
-            )
-            .subscribe();
+    delete(htmlElement: HTMLElement, salon: ISalon): void {
+        this.confirmDialogService.delete(htmlElement, 'salon.delete.question', {id: salon.place}).pipe(
+            filter(confirmed => confirmed),
+            switchMap(() => this.salonService.delete(salon.id)),
+            tap(() => this.load()), // Recharge les données
+        ).subscribe();
     }
 
     load(): void {
@@ -81,10 +78,5 @@ export class SalonComponent implements OnInit {
 
     previousState(): void {
         window.history.back();
-    }
-
-    refreshSalons(event: PaginationEvent): void {
-        this.salonsPaginated = this.salons.slice((event.page - 1) * event.pageSize,
-            (event.page - 1) * event.pageSize + event.pageSize);
     }
 }
